260311_production_recommendation/backend/requirements.txt
fastapi~=0.115.0
uvicorn~=0.34.0
python-dotenv>=1.2.0
openai>=1.12.0
azure-identity>=1.15.0
httpx>=0.26.0
pydantic>=2.5.0
numpy>=1.26.0
